date filtering was added
